VCH Creation API: Stop writing key files to disk
When creating a VCH via the API, we should not write key/certificate
files to the API server's disk.

This change introduces the behavior as a flag so that it could be
exposed as an option for the CLI at some point in the future, but
does not expose it at this time.