Use nanoseconds in container create timestamps (#6781)
This change stores container create timestamps in nanoseconds in the
portlayer instead of seconds like we did previously. This allows us to
store and sort container list output more granularly for the docker
personality.

Before this change, we were seeing failures where the 'docker ps -l'
command showed a container that wasn't the last one created. A possible
reason is that multiple containers can have the same second in their
create timestamps, and Go's sort.Sort is unstable w.r.t. the order of
equal values in a slice.

This commit includes:

  * using nsec instead of sec for create timestamps,
  * converting the stored timestamps to seconds in the docker persona
    for 'docker ps' and 'docker inspect', and
  * defining a new container data version so containers from older VCHs
    that don't use nsec timestamps still show valid ps and inspect
    output by multiplying sec timestamps by 1e9.

Fixes #6676