Put nfs nightly test infra deployment on the same pod (#6802)
All of the nfs infra should now deploy to the same POD.
That POD is dictated by the initial placement of the esxi
used for VCH deployment. If that fails the suite setup 
will be reattempted again a number of times after a 
cleanup process. This should stabilize the nfs nightlies 
more.