Fix docker pull on artifactory (#6831)
Artifactory responds with HTTP 403 when attempting initial query
to look up bearer realm.  We use an URL that includes manifest and
repo.  That works on most registries, and appears to be correct
according to the docker OAuth spec.

Now, if the URL we use does not work, we attempt another query
with just the registry URL.  This works well with Artifactory.  It
is also the same URL we use for docker login.

Fixes #6738