Define test plan for existing VCH Management APIs (#7501)
For 1.3.0, an initial set of functionality was implemented for the
vic-machine-service. Testing of that functionality was not systematic,
and was focused on the functionality that would be used by the UI.

Define a comprehensive test plan to ensure coverage and identify gaps.

Detailed test steps are not provided at this stage; these will be
added when the tests are implemented.