validator: Reduce visibility of methods (#7859)
Reduce visibility of validator methods used only within the package.

By limiting the package's public contract, it becomes easier to
understand how to use it as intended. Additionally, this allows for
more aggressive refactoring of such methods by making it clear that
they are internal implementation details of the package.