Update drone config to specify downstream branch (#8058)
The master branch of vic-product consumes the latest build from the
master branch of vic. Therefore it makes sense to trigger a build of
master for each push to master.

The release branches of vic-product only consume tag builds of vic.
Therefore it makes sense to only trigger new builds of vic-product's
release branches when performing a tag build of vic.