Additional fixes to integration tests (#8069)
This builds on #8047 and #8063 to address some outstanding issues left by
that work. In the case of the hetrogeneous and NFS tests, it is test
updates necessitated by the prior Nimbus-Util.robot updates.

In the case of the VSAN tests it switches to using a testbed descriptor
based off the vic-simple-cluster spec instead of what appeared to be a
library spec, testng/testbeds/vsan.rb.

This disables the complex vSAN testsuite at this time as I could not infer
what the complex case was testing, and it is not in the test doc. It's
possible that it's testing multiple vSAN datastores and our ability to use
either but that's pure speculation.