Temp setting of max_map_count in bootstrap (#7790)
The correct approach for this issue is to allow specification of sysctl
options generally and then parse/apply them in the container. This
change is made because:
a. it's extremely quick to do
b. it impacts a common workload (elastic search)
c. after consultation with Photon kernel team there's no known
negative impact to changing the default given cVM model.