Add per-field suppression to extraconfig encoder (#8038)
Adds a recursion property to skip encode or decode of a field. If
skipping both the depth=0 should be used.

Splits DefaultGuestInfoPrefix. It should have only contained the vSphere
mandatory guestinfo prefix but also contained the default VIC "vice"
namespace prefix. As part of this it extracts most string constants used
in the package and turns them into actual constants, and moves those
constants that should be modifiable to be variables. Future work should
extend this to be a defined config structure that Encode and Decode
can operate from.

Updates extraconfig to have a CalculateKey function in addition to the
CalculateKeys function that returns an array. This has been done because
of an observation that by far the most common usage was
CalculateKeys(...)[0], sometimes with length checking of the return but
frequently without. The new method panics if the structure pattern for
the field cannot be found.