Workaround for concurrent exec issues (#8098)
There are outstanding issues to address with concurrent exec. This work is
palliative rather than an actual fix.

Removes checking for "started" in the status string - we reliably see this
field not propagating to the property collector despite being logged as
set in the tether. This _only_ applies to execs at this time as that is
the only path calling task.State (via InspectTask).

Adds locking around dispatch of execs, with a timeout, to serialize that
initial dispatch path against a single container. If the timeout expires
it reverts to current behaviour and relies on concurrent modification and
retry.