Pprof server waits for required interface rename
The interfaces within the endpointVM get renamed to their network roles
and an entry is added to /etc/hosts so that it's easy for services to bind
to specific network roles irrespective of actual network configuration.
As components do not launch until the network is fully configured and
initialized this works as expected, but fails for the pprof server
embedded in vic-init that is performing the network initialization.

This adds logic to block the pprof initialization until the expected
name (client.localhost) resolves. This only impacts execution with debug
level of >=2.