Fix issues during container VM bootstrap
Delete the /etc/localtime file that shows in centos:6.9 based
bootstrap, causing the system clock to be set to EDT, but interpreted as
UTC after the switch to the container filesystem.

Stop the udev daemon after hardware has been initialized and before
switching to the container filesystem. This is done because the udev rules
will almost certainly not be present after the switch_root.
If hotadd is ever required then the rules will have to be added to
/.tether and udev configured to read from that tree while managing /dev.

Restore the important vsyscall option to the bootloader kernel
options, along with the others that were dropped.
Adds rgnd to the tmpfs size calculation.

Allow a repo config to specify specific kernel modules to be loaded
during the existing modprobe pass.
This was done to allow for modules that do not exist or have had name
changes between different kernel versions/vendors.

Also adds explicit load of xt_state module for iptables.

For some reason busybox based images were not correctly demand loading
xt_state module. All other images seemed to have no problem.
The failure was resulting in a failure to enable iptables configuration
for established connection tracking (which is the first rule applied to
the VIC chain and uses the xt_state module) and that prevented any
subsequent rules from being set, resulting in no outbound rule. This is
seen as "operation not permtted" error from ping, as an example.

Bring up loopback interface since systemd is not called for init. Package
net-tools conflicts with toybox so use ifconfig to bring up lo.

The tether explicitly logs to both stdout/stderr and to the serial debug
port. This change redirects stdout/stderr to the console so that we do not
end up with duplicate lines in the serial log. The stdout/stderr are
redirected to the serial log directly during bootstrap in order to log the
initializtion script output before the switch_root to tether.

The udevd in centos-6.9 (other distros unconfirmed) does not create the
symlink /dev/fd->/proc/self/fd. This results in various images (postgres
specifically) reporting errors such as:
initdb: could not open file "/dev/fd/63" for reading: No such file or
directory

The /dev/pts entries should be owned by group "tty" (gid=5) as has been
the convention for many years.
Likewise /dev/pts/ptmx should be mode 666 to allow creation of additional
pty device pairs.

Mount cgroups to support DCH. In some distro with older kernel version,
like centos6, cgroup is not mounted on /sys/fs/cgroup but on /cgroup,
in which case, we would not use cgroup. Now we test this directory's
existence before mounting it to prevent error message.