Avoid errors: ip conflict and ip address already in use (#8445)
If the container VM is operated from vSphere client, in some cases
VIC will allocate duplicate IP address for cVMs. For example: when
the first cVM is poweroff from VC, the assigned IP address will be
released, but the guestinfo is still saving the assigned IP address
for VC HA or migration functionality; the second cVM will be assigned
the first cVM ip address, when the first cVM is powered on from VC,
these two cVMs are using same IP address.

Another related error scenario is the docker network inspect will not
show container endpoint if cVM is powered on from vSphere.